[CSL-2209] Remove 'accepted clients' from CLI and server-side check
As discussed, this is only cluttering the feature at the moment. We don't do
anything with this and only have one client anyway. We may introduce that later on
if required.
Instead, I've replaced the check on the client's name with a check on
the keyusage purpose of the underlying client certificate since it was
the original intent with that name check (to prevent server certs to be
used instead of client certs).